feat(front): generate model with sw2dts
