+++ front/src/model.d.ts
@@ -1,20 +1,3 @@
-// this file generated automatically. DO NOT MODIFY.
+export * from './model-generated';
 
-export interface Noop {
-}
-export interface Todo {
-    createdAt?: string; // date-time
-    done?: boolean;
-    id?: string; // int64
-    text?: string;
-    updatedAt?: string; // date-time
-}
-export interface TodoListResp {
-    cursor?: string;
-    list?: Todo[];
-}
-export interface UserWelcomeResponse {
-    loggedIn?: boolean;
-    loginURL?: string;
-    logoutURL?: string;
-}
+// Add othrer types bellow
